requirements.txt
Package                   Version
------------------------- -----------
altair                    5.5.0
attrs                     25.3.0
black                     25.1.0
blinker                   1.9.0
cachetools                5.5.2
certifi                   2025.4.26
charset-normalizer        3.4.2
click                     8.2.1
gitdb                     4.0.12
GitPython                 3.1.44
greenlet                  3.2.3
idna                      3.10
Jinja2                    3.1.6
jsonschema                4.24.0
jsonschema-specifications 2025.4.1
MarkupSafe                3.0.2
mypy_extensions           1.1.0
narwhals                  1.42.1
numpy                     2.2.6
packaging                 24.2
pandas                    2.3.0
pathspec                  0.12.1
pillow                    11.2.1
pip                       22.0.2
platformdirs              4.3.8
protobuf                  6.31.1
psycopg2-binary           2.9.10
pyarrow                   20.0.0
pydeck                    0.9.1
python-dateutil           2.9.0.post0
python-dotenv             1.1.0
pytz                      2025.2
referencing               0.36.2
requests                  2.32.4
rpds-py                   0.25.1
setuptools                59.6.0
six                       1.17.0
smmap                     5.0.2
SQLAlchemy                2.0.41
streamlit                 1.45.1
tenacity                  9.1.2
toml                      0.10.2
tomli                     2.2.1
tornado                   6.5.1
typing_extensions         4.14.0
tzdata                    2025.2
urllib3                   2.4.0
watchdog                  6.0.0
